# save model checkpoint + run validation