--- a/tests/test_train_model.py
+++ b/tests/test_train_model.py
@@ -212,7 +212,7 @@ class TestTrainModel(unittest.TestCase):
             valid, _ = tl.train()
 
         self.assertEqual(tl.valid_reports[-1]['total_train_updates'], num_train_steps)
-        self.assertEqual(len(tl.valid_reports), num_validations)
+        self.assertEqual(len(tl.valid_reports), num_validations + 1)
         self.assertEqual(len(tl.train_reports), num_logs)  # log every valid as well
 
     def test_opt_step(self):